Update behavior.php
Second parameter in [script] function should be now [false] for [tabstate] function. This is because this tabs-state.js requires JQuery not full Mootools framework. Setting this to [true] not only load Mootools when it is not required but also cause some conflicts if websites that depends on JQuery without noConflict.